Finding contexts
Contexts are pairs of words w1 w2 with a word v1 between them, which would give w1 v1 w2. For example, 
"Some of the", where the pair is "Some __ The", and the word between is of. For each context like "Some __ the", I have written a program to see what are the most commonly occurring words.